refactor: extract shared ContentBlockedFetchError and handler utility
Architecture review (PR #861) identified two issues:

1. The isContentBlocked sentinel property on anonymous Error objects used
   unsafe intersection casts — replaced with a proper ContentBlockedFetchError
   class that supports clean instanceof checks.

2. The ~30-line CONTENT_BLOCKED handling block was duplicated verbatim in
   both nexus/page.tsx and decision-capture/page.tsx — extracted to
   lib/nexus/content-blocked-handler.ts.

The new handleContentBlockedResponse(response, log) utility:
- Checks for 400 status with CONTENT_BLOCKED code
- Shows toast with blocked categories
- Logs the event with structured data
- Throws ContentBlockedFetchError (caught by instanceof, not string matching)

Both page components now call a single line instead of 30 lines of
inline error handling.

Part of #860
